+++ CHANGELOG.md
@@ -4,10 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Improved
+- Remote webdriver session ID is now logged after connecting #TINY-10835
+
+## Fixed
+- Failing tests that retry and then pass could exceed the "tests crashed" server timeout #TINY-11177
+
 ## 15.0.1 - 2025-01-27
 
 ## Fixed
-- AWS remote testing idle timeout was not set
+- AWS remote testing idle timeout was not set #TINY-11177
 
 ## 15.0.0 - 2025-01-27